+++ Tabloid/client/src/providers/CommentProvider.js
@@ -25,7 +25,30 @@ export const CommentProvider = (props) => {
     fetch(`${apiUrl}`, {
       method: "POST",
       headers: {
-        "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
+      },
+      body: JSON.stringify(comment),
+    })
+    .then((res) => res.json()));
+  }
+
+  const getCommentById = (id) => {
+    return getToken().then((token) =>
+    fetch(`/api/Comment/GetCommentById/${id}`, {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      }
+    })
+      .then((res) => res.json()));
+  }  
+
+  const updateComment = (comment) => {
+    return getToken().then((token) => 
+    fetch(`/api/Comment/${comment.id}`, {
+      method: "PUT",
+      headers: {
+        Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify(comment),
     })
@@ -33,7 +56,7 @@ export const CommentProvider = (props) => {
   }
 
   return (
-    <CommentContext.Provider value={{ comments, setComments, postId, setPostId, getAllCommentsOnPost, addComment }}>
+    <CommentContext.Provider value={{ comments, setComments, postId, setPostId, getAllCommentsOnPost, addComment, getCommentById, updateComment }}>
       {props.children}
     </CommentContext.Provider>
   )
